change remaining_finals storage type to uint32_t

+++ include/zenoh-pico/session/session.h
@@ -194,7 +194,7 @@ struct _z_pending_query_t {
     z_clock_t _start_time;
     uint64_t _timeout;
     void *_arg;
-    uint8_t _remaining_finals;
+    uint32_t _remaining_finals;
     _z_pending_reply_slist_t *_pending_replies;
     z_query_target_t _target;
     z_consolidation_mode_t _consolidation;
